New HDF5 file format
The Multivariate emulator has now a new HDF5 file format. The idea
is that all the emulators for a particular model (the angular
library) are stored in the same file. There are a couple of missing
things yet. These are:

1. A convenient automatic way of informing the library where all the
emulators are stored (e.g. an environment variable), so that one
could query the available emulators for example.
2. A possibility to e.g. query the spectral range of the emulator
3. We now have a new dependency: we need h5py