Added an implementation of the Hessian to the multivariate GP emulator code.
The implementation is based on calculating the individual Hessian
for each of the weight factors, multiplying them by the relevant
basis function, and adding them together. The output array is thus
(Npar, Npar, len(basis_functions)) (e.g. (10, 10, 2101) for PROSAIL).

I haven't written any tests to test this, but a simple test would be
to train a single band emulator, and calculate the Hessian for a
particular input test vector. Then calculate the Hessian for the
multivariate emulator, and integrate over the passband of the single
band emulator. The resultint two 10*10 matrices should be identical.