Removing dependencies on h5py
In some installations, the dependency on h5py (HDF5) results in
the Python interpreter bombing if the h5py library has been compiled
with different library version to the one being used in the current
system. Since I never really got around to saving stuff as HDF5
anyway, I switched it off as a dependency.