Added consistency in predict method
Now both the multivariate and univarate emulators allow you to
switch the uncertainty and gradient calculations on or off. There's
a slight difference in the API: the predict methods return three
parameters irrespective of `do_unc` or `do_deriv` flags. Only that
in some cases, they'll be `None` (if the required option was set
to `False`). This might need some changes to code that depends on
`gp_emulator`.

Also, haven't tested the changes yet!